partially works
this program is now be able to run
added Qimg2cv and cv2Qimg, make it be able to convert from both side.

Houghcirle is not functioning well, it cannot detected circle properly. need to work on change parameters to see if works.
be awear of line 177
or may have to do some pre-process to make it work
also, frame rate dropping as well.