problems for perfect hit bars
these two functions need to be rewrite, not the whole thing,
for set transforms function, transform matries need to be redefine, we have to find out which numbers are affecting the wrist movement, and fix those numbers to make sure it always back to the proper position for hitting then we can use the original userhitting function.
if we don't change the transform function, we may have to redefine those hitting parameters in user hit function, change the returning angles would be the best way to go for this function, however, it might not that easy.
these are two ways of making hit perfect that I can think of for now.