creating control module for the whole session
this class/module is to have the main control of one big session.
in this class will have a flow control which tells robot which task to do, basicly having the whole session under contorl.
may have think about it very careful!!
should have the ability to stop the whole session in case of emergency or be able to start from the stopping point if session gets continued.