try to create a tempreture control
it looks like not that important for now, if temp goes high I still cannot unlock joints, otherwise it will effect the inaccuracy for play music
however, I may have to include some of the functions like enablenotification to turn of the warning message. somewhere in main function
may have to delete this file, keep it for now.