adding csv data saving
now adding save csv data to the code
however, this may not work properly
have to test it and see how it runs
also need to test the audio processing part, in order to process the notes which recorded from the robot., not sure if that can be perfectly done, if not, have to re modify the notes frequencies and other datas.
also have to determine when bad hit happens, how can we handle that part. including low volume inputs, mute kind hit, hit too strong, hit other places, not clear sounds and etc.,
these should be done over the weekend, and have to record a video in order to show the whole process to Anible.